[#36][news] refactor: 불필요 로직 삭제, 변수명 변경, 메서드명 변경
