+++ news/src/main/java/com/newpick4u/news/news/application/usecase/NewsService.java
@@ -1,9 +1,9 @@
 package com.newpick4u.news.news.application.usecase;
 
-import com.newpick4u.news.news.infrastructure.kafka.dto.NewsInfoDto;
-import com.newpick4u.news.news.infrastructure.kafka.dto.NewsTagDto;
+import com.newpick4u.news.news.application.dto.NewsInfoDto;
+import com.newpick4u.news.news.application.dto.NewsTagDto;
 
 public interface NewsService {
-    void createNewsInfo(NewsInfoDto dto);
-    void updateTagList(NewsTagDto dto);
-    }
+    void saveNewsInfo(NewsInfoDto dto);
+    void updateNewsTagList(NewsTagDto dto);
+}
