+++ news/src/main/java/com/newpick4u/news/news/application/usecase/TagInboxProcessor.java
@@ -1,10 +1,10 @@
 package com.newpick4u.news.news.application.usecase;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.newpick4u.news.news.application.dto.NewsTagDto;
 import com.newpick4u.news.news.domain.entity.News;
 import com.newpick4u.news.news.domain.entity.NewsTag;
 import com.newpick4u.news.news.domain.entity.TagInbox;
-import com.newpick4u.news.news.infrastructure.kafka.dto.NewsTagDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
@@ -24,7 +24,7 @@ public class TagInboxProcessor {
                     .map(tag -> NewsTag.create(tag.id(), tag.name(), news))
                     .toList();
 
-            news.updateNewsTags(tags);
+            news.addTags(tags);
         } catch (Exception e) {
             throw new RuntimeException("인박스 메시지 처리 실패", e);
         }
