+++ news/src/main/java/com/newpick4u/news/news/infrastructure/kafka/NewsTagConsumer.java
@@ -1,7 +1,7 @@
 package com.newpick4u.news.news.infrastructure.kafka;
 
-import com.newpick4u.news.news.application.usecase.NewsMessageHandler;
-import com.newpick4u.news.news.infrastructure.kafka.dto.NewsTagDto;
+import com.newpick4u.news.news.application.dto.NewsTagDto;
+import com.newpick4u.news.news.application.usecase.NewsService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
@@ -14,16 +14,16 @@ import org.springframework.stereotype.Component;
 @RequiredArgsConstructor
 public class NewsTagConsumer {
 
-    private final NewsMessageHandler newsMessageHandler;
+    private final NewsService newsService;
 
     @KafkaListener(
-            topics = "tag.fct.1",
+            topics = "tag.fct.v1",
             groupId = "news-tag-consumer",
             containerFactory = "newsTagListenerContainerFactory"
     )
     public void consume(ConsumerRecord<String, NewsTagDto> record, Acknowledgment ack) {
         try {
-            newsMessageHandler.handleNewsTagUpdate(record.value());
+            newsService.updateNewsTagList(record.value());
             ack.acknowledge();
         } catch (Exception e) {
             log.error("[Kafka] 뉴스 태그 연동 실패: {}", record.value(), e);
