+++ news/src/main/java/com/newpick4u/news/news/infrastructure/kafka/RetryNewsTagConsumer.java
@@ -1,7 +1,7 @@
 package com.newpick4u.news.news.infrastructure.kafka;
 
-import com.newpick4u.news.news.application.usecase.NewsMessageHandler;
-import com.newpick4u.news.news.infrastructure.kafka.dto.NewsTagDto;
+import com.newpick4u.news.news.application.dto.NewsTagDto;
+import com.newpick4u.news.news.application.usecase.NewsService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
@@ -13,10 +13,10 @@ import org.springframework.stereotype.Component;
 @RequiredArgsConstructor
 public class RetryNewsTagConsumer {
 
-    private final NewsMessageHandler newsMessageHandler;
+    private final NewsService newsService;
 
     @KafkaListener(
-            topics = "tag.dlq.v1",
+            topics = "tag-dlq.fct.v1",
             groupId = "news-tag-dlq-consumer",
             containerFactory = "newsTagListenerContainerFactory"
     )
@@ -24,7 +24,7 @@ public class RetryNewsTagConsumer {
         NewsTagDto dto = record.value();
         try {
             log.info("[DLQ Retry] 태그 DLQ 재처리 시작: {}", dto);
-            newsMessageHandler.handleNewsTagUpdate(dto);
+            newsService.updateNewsTagList(dto);
         } catch (Exception e) {
             log.error("[DLQ Retry] 태그 재처리 실패 - 메시지: {}", dto, e);
             throw new RuntimeException(e);
